Update utils structure
