+++ packages/shopify/src/cart/use-update-item.tsx
@@ -1,25 +1,27 @@
 import { useCallback } from 'react'
 import debounce from 'lodash.debounce'
-import type {
-  HookFetcherContext,
-  MutationHookContext,
-} from '@vercel/commerce/utils/types'
 import { ValidationError } from '@vercel/commerce/utils/errors'
-import useUpdateItem, {
-  UseUpdateItem,
-} from '@vercel/commerce/cart/use-update-item'
 
 import useCart from './use-cart'
-import { handler as removeItemHandler } from './use-remove-item'
-
-import { getCartId, normalizeCart, cartLinesUpdateMutation } from '../utils'
-
-import type { UpdateItemHook, LineItem } from '../types/cart'
 
 import type {
   CartLinesUpdateMutation,
   CartLinesUpdateMutationVariables,
 } from '../../schema'
+
+import type {
+  HookFetcherContext,
+  MutationHookContext,
+} from '@vercel/commerce/utils/types'
+
+import type { UpdateItemHook, LineItem } from '../types/cart'
+
+import useUpdateItem, {
+  UseUpdateItem,
+} from '@vercel/commerce/cart/use-update-item'
+
+import { handler as removeItemHandler } from './use-remove-item'
+import { getCartId, normalizeCart, cartLinesUpdateMutation } from '../utils'
 
 export type UpdateItemActionInput<T = any> = T extends LineItem
   ? Partial<UpdateItemHook['actionInput']>
