+++ packages/shopify/src/fetcher.ts
@@ -1,6 +1,6 @@
 import { Fetcher } from '@vercel/commerce/utils/types'
 import { API_TOKEN, API_URL } from './const'
-import { handleFetchResponse } from './utils'
+import { handleFetchResponse } from './utils/handle-fetch-response'
 
 const fetcher: Fetcher = async ({
   url = API_URL,
