+++ packages/shopify/src/product/use-search.tsx
@@ -1,7 +1,7 @@
 import { SWRHook } from '@vercel/commerce/utils/types'
 import useSearch, { UseSearch } from '@vercel/commerce/product/use-search'
 
-import {
+import type {
   CollectionEdge,
   GetAllProductsQuery,
   GetProductsFromCollectionQueryVariables,
@@ -9,14 +9,14 @@ import {
   ProductEdge,
 } from '../../schema'
 
+import type { SearchProductsHook } from '../types/product'
+
 import {
   getAllProductsQuery,
   getCollectionProductsQuery,
   getSearchVariables,
   normalizeProduct,
 } from '../utils'
-
-import type { SearchProductsHook } from '../types/product'
 
 export type SearchProductsInput = {
   search?: string
