--- packages/shopify/src/utils/handle-account-activation.ts
@@ -1,28 +1,0 @@
-import { FetcherOptions } from '@vercel/commerce/utils/types'
-import { throwUserErrors } from './throw-user-errors'
-
-import {
-  MutationCustomerActivateArgs,
-  MutationCustomerActivateByUrlArgs,
-} from '../../schema'
-import { Mutation } from '../../schema'
-import { customerActivateByUrlMutation } from './mutations'
-
-export const handleAccountActivation = async (
-  fetch: <T = any, B = Body>(options: FetcherOptions<B>) => Promise<T>,
-  input: MutationCustomerActivateByUrlArgs
-) => {
-  try {
-    const { customerActivateByUrl } = await fetch<
-      Mutation,
-      MutationCustomerActivateArgs
-    >({
-      query: customerActivateByUrlMutation,
-      variables: {
-        input,
-      },
-    })
-
-    throwUserErrors(customerActivateByUrl?.customerUserErrors)
-  } catch (error) {}
-}
